templates/config-python/requirements.txt
# Agentic QE Framework v2 — Python dependencies
playwright>=1.50.0
pytest>=8.0.0
pytest-playwright>=0.5.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pytest-html>=4.0.0
